patch: refactor auth handling. seperate each auth method into their own module.
