+++ Minecraft.World/cmake/sources/Common.cmake
@@ -261,6 +261,10 @@ set(_MINECRAFT_WORLD_COMMON_NET_MINECRAFT_NETWORK_PACKET
   "${CMAKE_CURRENT_SOURCE_DIR}/AnimatePacket.h"
   "${CMAKE_CURRENT_SOURCE_DIR}/AuthModule.cpp"
   "${CMAKE_CURRENT_SOURCE_DIR}/AuthModule.h"
+  "${CMAKE_CURRENT_SOURCE_DIR}/SessionAuthModule.cpp"
+  "${CMAKE_CURRENT_SOURCE_DIR}/SessionAuthModule.h"
+  "${CMAKE_CURRENT_SOURCE_DIR}/OfflineAuthModule.cpp"
+  "${CMAKE_CURRENT_SOURCE_DIR}/OfflineAuthModule.h"
   "${CMAKE_CURRENT_SOURCE_DIR}/AuthPackets.cpp"
   "${CMAKE_CURRENT_SOURCE_DIR}/AuthPackets.h"
   "${CMAKE_CURRENT_SOURCE_DIR}/UUID.cpp"
